+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Memory.Box/UX/KryptonMemoryBoxOld.resx
